+++ lib/aj_image_editor.dart
@@ -1,3 +1,5 @@
 library aj_image_editor;
 
 export 'src/image_editor.dart';
+export 'src/Enum/image_type.dart';
+export 'src/Enum/shape_type.dart';